requirements.txt
flask>=2.3.0
tensorflow>=2.10.0
pillow>=10.0.0
numpy>=1.23.0
scikit-learn>=1.3.0
